requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
requests
docxtpl
pandas
openpyxl